refactor(plugin-lighthouse): move runner logic into separate folder
