+++ packages/plugin-lighthouse/src/index.ts
@@ -1,11 +1,7 @@
 import { lighthousePlugin } from './lib/lighthouse-plugin';
 
-export { lighthousePlugin, LighthouseCliFlags } from './lib/lighthouse-plugin';
-export {
-  LIGHTHOUSE_REPORT_NAME,
-  LIGHTHOUSE_PLUGIN_SLUG,
-  LIGHTHOUSE_AUDITS,
-  LIGHTHOUSE_GROUPS,
-} from './lib/constants';
-
 export default lighthousePlugin;
+export { lighthousePlugin } from './lib/lighthouse-plugin';
+export { LIGHTHOUSE_PLUGIN_SLUG } from './lib/constants';
+export * from './lib/runner';
+export { lighthouseAuditRef, lighthouseGroupRef } from './lib/utils';
