+++ packages/plugin-lighthouse/src/lib/runner/index.ts
@@ -1,0 +1,7 @@
+export { createRunnerFunction, LighthouseCliFlags } from './runner';
+export {
+  LIGHTHOUSE_REPORT_NAME,
+  LIGHTHOUSE_NAVIGATION_AUDITS,
+  LIGHTHOUSE_GROUPS,
+  LighthouseGroupSlugs,
+} from './constants';
