+++ packages/plugin-lighthouse/src/lib/utils.ts
@@ -1,69 +1,37 @@
-import chalk from 'chalk';
-import { type Budget, type CliFlags, type Config } from 'lighthouse';
-import log from 'lighthouse-logger';
-import desktopConfig from 'lighthouse/core/config/desktop-config.js';
-import experimentalConfig from 'lighthouse/core/config/experimental-config.js';
-import perfConfig from 'lighthouse/core/config/perf-config.js';
-import { Result } from 'lighthouse/types/lhr/audit-result';
-import path from 'node:path';
-import { Audit, AuditOutput, AuditOutputs, Group } from '@code-pushup/models';
-import {
-  filterItemRefsBy,
-  importEsmModule,
-  objectToCliArgs,
-  readJsonFile,
-  toArray,
-  ui,
-} from '@code-pushup/utils';
-import { LIGHTHOUSE_REPORT_NAME } from './constants';
-import { type LighthouseCliFlags } from './lighthouse-plugin';
+import { type CliFlags } from 'lighthouse';
+import { Audit, CategoryRef, Group } from '@code-pushup/models';
+import { filterItemRefsBy, toArray } from '@code-pushup/utils';
+import { LIGHTHOUSE_PLUGIN_SLUG } from './constants';
+import { LighthouseGroupSlugs } from './runner';
+
+export function lighthouseGroupRef(
+  groupSlug: LighthouseGroupSlugs,
+  weight = 1,
+): CategoryRef {
+  return {
+    plugin: LIGHTHOUSE_PLUGIN_SLUG,
+    slug: groupSlug,
+    type: 'group',
+    weight,
+  };
+}
+
+export function lighthouseAuditRef(auditSlug: string, weight = 1): CategoryRef {
+  return {
+    plugin: LIGHTHOUSE_PLUGIN_SLUG,
+    slug: auditSlug,
+    type: 'audit',
+    weight,
+  };
+}
 
 type RefinedLighthouseOption = {
   url: CliFlags['_'];
   chromeFlags?: Record<CliFlags['chromeFlags'][number], string>;
 };
+
 export type LighthouseCliOptions = RefinedLighthouseOption &
   Partial<Omit<CliFlags, keyof RefinedLighthouseOption>>;
-
-export function getLighthouseCliArguments(
-  options: LighthouseCliOptions,
-): string[] {
-  const {
-    url,
-    outputPath = LIGHTHOUSE_REPORT_NAME,
-    onlyAudits = [],
-    output = 'json',
-    verbose = false,
-    chromeFlags = {},
-  } = options;
-
-  // eslint-disable-next-line functional/no-let
-  let argsObj: Record<string, unknown> = {
-    _: ['lighthouse', url.join(',')],
-    verbose,
-    output,
-    'output-path': outputPath,
-  };
-
-  if (onlyAudits != null && onlyAudits.length > 0) {
-    argsObj = {
-      ...argsObj,
-      onlyAudits,
-    };
-  }
-
-  // handle chrome flags
-  if (Object.keys(chromeFlags).length > 0) {
-    argsObj = {
-      ...argsObj,
-      chromeFlags: Object.entries(chromeFlags)
-        .map(([key, value]) => `--${key}=${value}`)
-        .join(' '),
-    };
-  }
-
-  return objectToCliArgs(argsObj);
-}
 
 export class AuditsNotImplementedError extends Error {
   constructor(auditSlugs: string[]) {
@@ -82,37 +50,6 @@ export function validateOnlyAudits(
     throw new AuditsNotImplementedError(missingAudtis);
   }
   return true;
-}
-
-export function toAuditOutputs(lhrAudits: Result[]): AuditOutputs {
-  return lhrAudits.map(
-    ({
-      id: slug,
-      score,
-      numericValue: value = 0, // not every audit has a numericValue
-      details,
-      displayValue,
-    }: Result) => {
-      const auditOutput: AuditOutput = {
-        slug,
-        score: score ?? 1, // score can be null
-        value: Number.parseInt(value.toString(), 10),
-        displayValue,
-      };
-
-      if (details == null) {
-        return auditOutput;
-      }
-
-      // @TODO implement switch case for detail parsing. Related to #90
-      const unsupportedType = details.type;
-      ui().logger.info(
-        `Parsing details from type ${unsupportedType} is not implemented.`,
-      );
-
-      return auditOutput;
-    },
-  );
 }
 
 export class CategoriesNotImplementedError extends Error {
@@ -148,9 +85,9 @@ export function filterAuditsAndGroupsByOnlyOptions(
   if (onlyCategories && onlyCategories.length > 0) {
     validateOnlyCategories(groups, onlyCategories);
 
-    const categorieSlugs = new Set(onlyCategories);
+    const categorySlugs = new Set(onlyCategories);
     const filteredGroups: Group[] = groups.filter(({ slug }) =>
-      categorieSlugs.has(slug),
+      categorySlugs.has(slug),
     );
     const auditSlugsFromRemainingGroups = new Set(
       filteredGroups.flatMap(({ refs }) => refs.map(({ slug }) => slug)),
@@ -174,91 +111,4 @@ export function filterAuditsAndGroupsByOnlyOptions(
     audits,
     groups,
   };
-}
-
-export async function getConfig(
-  flags: Pick<LighthouseCliFlags, 'configPath' | 'preset'> = {},
-): Promise<Config | undefined> {
-  const { configPath: filepath, preset } = flags;
-
-  if (filepath != null) {
-    // Resolve the config file path relative to where cli was called.
-
-    if (filepath.endsWith('.json')) {
-      return readJsonFile<Config>(filepath);
-    } else if (/\.(ts|js|mjs)$/.test(filepath)) {
-      return importEsmModule<Config>({ filepath });
-    }
-  } else if (typeof preset === 'string') {
-    switch (preset) {
-      case 'desktop':
-        return desktopConfig;
-      case 'perf':
-        return perfConfig as Config;
-      case 'experimental':
-        return experimentalConfig as Config;
-      default:
-        // as preset is a string literal the default case here is normally caught by TS and not possible to happen. Now in reality it can happen and preset could be a string not included in the literal.
-        // Therefor we have to use `as string` is used. Otherwise, it will consider preset as type never
-        ui().logger.info(`Preset "${preset as string}" is not supported`);
-    }
-  }
-  return undefined;
-}
-
-export async function getBudgets(
-  budgetPath?: string | null,
-): Promise<Budget[]> {
-  if (budgetPath) {
-    /** @type {Array<LH.Budget>} */
-    return await readJsonFile<Budget[]>(
-      path.resolve(process.cwd(), budgetPath),
-    );
-  }
-  return [];
-}
-
-export function setLogLevel({
-  verbose,
-  quiet,
-}: {
-  verbose?: boolean;
-  quiet?: boolean;
-} = {}) {
-  // set logging preferences
-  if (verbose) {
-    log.setLevel('verbose');
-  } else if (quiet) {
-    log.setLevel('silent');
-  } else {
-    log.setLevel('info');
-  }
-}
-
-const excludedFlags = new Set([
-  // lighthouse CLI specific debug logs
-  'list-all-audits', // Prints a list of all available audits and exits.
-  'list-locales', // Prints a list of all supported locales and exits.
-  'list-trace-categories', // Prints a list of all required trace categories and exits.
-]);
-
-export function validateFlags(
-  flags: LighthouseCliFlags = {},
-): LighthouseCliFlags {
-  const unsupportedFlagsInUse = Object.keys(flags).filter(flag =>
-    excludedFlags.has(flag),
-  );
-
-  if (unsupportedFlagsInUse.length > 0) {
-    ui().logger.info(
-      `${chalk.yellow(
-        '⚠',
-      )} The following used flags are not supported: ${chalk.bold(
-        unsupportedFlagsInUse.join(', '),
-      )}`,
-    );
-  }
-  return Object.fromEntries(
-    Object.entries(flags).filter(([flagName]) => !excludedFlags.has(flagName)),
-  );
 }
